refactor: refactored verification component
