+++ packages/p2p/src/components/verification/index.ts
@@ -1,0 +1,4 @@
+import verification from './verification';
+import './verification.scss';
+
+export default verification;
