+++ packages/p2p/src/components/verification/verification.tsx
@@ -1,16 +1,18 @@
 import React from 'react';
 import classNames from 'classnames';
-import PropTypes from 'prop-types';
+import { observer } from 'mobx-react-lite';
 import { Icon, Checklist, Text } from '@deriv/components';
 import { isDesktop, isMobile, routes } from '@deriv/shared';
-import { observer } from 'mobx-react-lite';
-import { useStores } from 'Stores';
 import Dp2pBlocked from 'Components/dp2p-blocked';
 import { Localize } from 'Components/i18next';
 import { useModalManagerContext } from 'Components/modal-manager/modal-manager-context';
-import './verification.scss';
+import { useStores } from 'Stores/index';
 
-const VerificationWrapper = ({ should_wrap, children }) => {
+type TVerificationProps = {
+    should_wrap?: boolean;
+};
+
+const VerificationWrapper = ({ should_wrap = false, children }: React.PropsWithChildren<TVerificationProps>) => {
     if (should_wrap) {
         return (
             <div
@@ -24,10 +26,10 @@ const VerificationWrapper = ({ should_wrap, children }) => {
         );
     }
 
-    return children;
+    return <>{children}</>;
 };
 
-const Verification = ({ should_wrap }) => {
+const Verification = ({ should_wrap = false }: TVerificationProps) => {
     const { general_store } = useStores();
     const { showModal } = useModalManagerContext();
 
@@ -40,7 +42,9 @@ const Verification = ({ should_wrap }) => {
             content: general_store.nickname || <Localize i18n_default_text='Choose your nickname' />,
             status: general_store.nickname ? 'done' : 'action',
             onClick: general_store.nickname
-                ? () => {}
+                ? () => {
+                      //do nothing
+                  }
                 : () => {
                       if (isDesktop()) showModal({ key: 'NicknameModal' });
                       general_store.toggleNicknamePopup();
@@ -52,7 +56,9 @@ const Verification = ({ should_wrap }) => {
             status: general_store.poi_status === 'verified' ? 'done' : 'action',
             onClick:
                 general_store.poi_status === 'verified'
-                    ? () => {}
+                    ? () => {
+                          //do nothing
+                      }
                     : () => {
                           const search = window.location.search;
                           let updated_url = `${routes.proof_of_identity}?ext_platform_url=${routes.cashier_p2p}`;
@@ -85,10 +91,6 @@ const Verification = ({ should_wrap }) => {
             </div>
         </VerificationWrapper>
     );
-};
-
-Verification.propTypes = {
-    should_wrap: PropTypes.bool,
 };
 
 export default observer(Verification);
