+++ packages/p2p/src/pages/buy-sell/buy-sell.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { observer } from 'mobx-react-lite';
 import { localize } from 'Components/i18next';
 import PageReturn from 'Components/page-return';
-import Verification from 'Components/verification/verification.jsx';
+import Verification from 'Components/verification';
 import { buy_sell } from 'Constants/buy-sell';
 import { useStores } from 'Stores';
 import BuySellHeader from './buy-sell-header.jsx';
